requirements.txt
fastapi>=0.100.0
uvicorn>=0.20.0
openai>=1.0.0
requests>=2.25.0
gql>=3.4.0
python-dotenv>=0.19.0
pydantic>=2.0.0 